fix bugs cause by merge low memory

+++ include/type.h
@@ -3,7 +3,7 @@
 typedef unsigned char u8;
 typedef unsigned int  u32;
 typedef unsigned short u16;
-typedef unsigned long u64;
+typedef  u32 size_t;
 #define NULL 0
 #endif
